backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo==4.8.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-dotenv==1.0.1
pydantic[email]==2.8.2
httpx==0.27.2
pytz==2024.1
itsdangerous==2.2.0
